+++ R/4_2_wordclouds.R
@@ -582,7 +582,7 @@ create_plots <- function(
                               ggplot2::aes(label = ngrams, 
                                            size = prop, 
                                            color = prop)) +
-        ggwordcloud::geom_text_wordcloud() +
+        ggwordcloud::geom_text_wordcloud(show.legend = TRUE) +
         ggplot2::theme_minimal() +
         color_positive_cor # ggplot2::scale_color_gradient(low = "yellow", high = "red")
       
@@ -634,9 +634,9 @@ create_plots <- function(
           # Word cloud with correlation strength mapped to color gradient
           plot1 <- ggplot2::ggplot(test_positive, 
                                    ggplot2::aes(label = top_terms,
-                                                size = estimate,
-                                                color = prop)) +
-            ggwordcloud::geom_text_wordcloud() +
+                                                size = prop,
+                                                color = estimate)) +
+            ggwordcloud::geom_text_wordcloud(show.legend = TRUE) +
             ggplot2::scale_size_area(max_size = max_size) +  # Adjust max size
             #scale_color_gradient(low = "grey", high = "red") +  # Blue for low, red for high correlation strength
             ggplot2::theme_minimal() +